# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: StreamGate API
  description: |
    NFT-gated streaming platform API. Core flow: wallet sign-in → NFT verification → protected HLS streaming.

    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## NFT Permission Control
    Content is protected by NFT ownership. Users must hold the specified NFT to access protected streaming content.
  version: 1.0.0
  contact:
    name: StreamGate Team
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /api/v1
    description: Relative path (behind proxy)
  - url: http://localhost:8080/api/v1
    description: Local development
...
paths:
...
  /auth/challenge:
    post:
      tags: [Auth]
      summary: Request a sign-in challenge
      description: Returns a nonce for the wallet to sign (EIP-191/EIP-712/SIWE)
      operationId: postAuthChallenge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChallengeRequest"
      responses:
        "200":
          description: Challenge created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChallengeResponse"
  /auth/login:
    post:
      tags: [Auth]
      summary: Wallet login
      description: Verifies wallet signature and returns JWT
      operationId: postAuthLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid signature
  /auth/register:
    post:
      tags: [Auth]
      summary: Register new user
      description: Registers a new user with username and password
      operationId: postAuthRegister
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Registration successful
        "409":
          description: User already exists
  /auth/verify:
    post:
      tags: [Auth]
      summary: Verify JWT token
      description: Validates a JWT token and returns its claims
      operationId: postAuthVerify
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyTokenRequest"
      responses:
        "200":
          description: Token is valid
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh JWT token
      operationId: postAuthRefresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout
      description: Revokes the current JWT token
      operationId: postAuthLogout
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out
  /auth/profile:
    get:
      tags: [Auth]
      summary: Get user profile
      description: Returns the authenticated user's profile including wallet address
      operationId: getAuthProfile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileResponse"
        "401":
          description: Unauthorized
  /auth/change-password:
    post:
      tags: [Auth]
      summary: Change password
      description: Changes password for username-password users (not available for wallet users)
      operationId: postChangePassword
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password changed
        "400":
          description: Wallet users cannot change password
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ChallengeRequest:
      type: object
      required: [wallet_address]
      properties:
        wallet_address:
          type: string
          description: Wallet address to generate challenge for
          example: "0x1234567890123456789012345678901234567890"
    ChallengeResponse:
      type: object
      properties:
        nonce:
          type: string
          description: Challenge nonce to sign
          example: "abc123def456"
        message:
          type: string
          description: Full message to sign (EIP-191/EIP-712/SIWE)
          example: "Sign this message to authenticate: abc123def456"
        expires_at:
          type: string
          format: date-time
    LoginRequest:
      type: object
      required: [wallet_address, signature]
      properties:
        wallet_address:
          type: string
        signature:
          type: string
          description: EIP-191 signed message
    RegisterRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
          format: password
    VerifyTokenRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
          description: JWT token to verify
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT access token
        refresh_token:
          type: string
        expires_at:
          type: string
          format: date-time
    ProfileResponse:
      type: object
      properties:
        wallet_address:
          type: string
        username:
          type: string
        created_at:
          type: string
          format: date-time
    ChangePasswordRequest:
      type: object
      required: [old_password, new_password]
      properties:
        old_password:
          type: string
          format: password
        new_password:
          type: string
          format: password
...
security:
  - bearerAuth: []